updated wine to ubuntu
Former-commit-id: b455beed7d152a82b5992f7a62ecaeb0f862c15f

+++ build_all.sh
@@ -13,6 +13,8 @@ build_and_push(){
 	suite=$2
 	build_dir=$3
 
+	[ -f "${build_dir}/.skip_buiild" ] && return 0;
+
 	echo "Building ${REPO_URL}/${base}:${suite} for context ${build_dir}"
 	docker build --rm --force-rm -t ${REPO_URL}/${base}:${suite} ${build_dir} || return 1
